test: 🧪 ParallelPop
